Added live search
