--- src/common/components/index.tsx
@@ -1,6 +1,0 @@
-import DefaultLayout from './layout/DefaultLayout';
-import { ComponentType, PropsType } from './types';
-import Component from './Component';
-import Tree from './Tree';
-
-export { DefaultLayout, ComponentType, PropsType, Component, Tree }
