--- src/common/components/types.ts
@@ -1,12 +1,0 @@
-export interface ComponentType {
-    srcPath: string;
-    className: string;
-    description?: string;
-    props?: PropsType[];
-}
-
-export interface PropsType {
-    name: string;
-    type: string;
-    description?: string;
-}
