+++ src/common/index.tsx
@@ -1,0 +1,5 @@
+import { ComponentType, PropsType } from './types';
+import Component from './Component';
+import Tree from './Tree';
+
+export { ComponentType, PropsType, Component, Tree }
